[BACKLOG-5807] Port the bar chart to VizAPI 3.0 (minimum set of features for demo) * Renamed Chart files to View
