+++ package-res/resources/web/pentaho/visual/ccc/barAbstract/model.js
@@ -20,7 +20,6 @@ define([
 ], function(visualFactory, bundle) {
 
   "use strict";
-
   
   return function(context) {
