+++ package-res/resources/web/pentaho/visual/ccc/boxplot/View.js
@@ -14,7 +14,7 @@
 * limitations under the License.
 */
 define([
-  "../categoricalContinuousAbstract/AbstractCategoricalContinuousChart"
+  "../categoricalContinuousAbstract/View"
 ], function(AbstractCategoricalContinuousChart) {
 
   "use strict";
