+++ package-res/resources/web/pentaho/visual/ccc/sunburst/View.js
@@ -16,7 +16,7 @@
 define([
   "cdf/lib/CCC/protovis",
   "cdf/lib/CCC/def",
-  "../abstract/AbstractChart",
+  "../abstract/View",
   "../util"
 ], function(pv, def, AbstractChart, util) {
